Relocated APIKey logic
